devops/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-agent"
version = "1.0.0"
description = "Agentic DevOps Automation - AI-powered CI/CD pipeline generation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "STRIDE Team"},
]
keywords = ["devops", "automation", "gemini", "ai", "cicd", "docker", "terraform"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "google-generativeai>=0.8.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "docker>=7.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "aiofiles>=23.0",
    "structlog>=24.0",
    "rich>=13.0",
    "typer>=0.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
devops-agent = "devops_agent.main:app"

[project.urls]
Homepage = "https://github.com/stride/devops-agent"
Documentation = "https://github.com/stride/devops-agent#readme"
Repository = "https://github.com/stride/devops-agent"

[tool.setuptools.packages.find]
where = ["."]
include = ["devops_agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
